Rename file mapping.py to types.py
