+++ ngraph_onnx/onnx_importer/ops_bridge/opset_01.py
@@ -32,7 +32,7 @@ from ngraph_onnx.onnx_importer.utils.binary import broadcast_for_binary_operatio
 from ngraph_onnx.onnx_importer.utils.conv import make_convolution_op
 from ngraph_onnx.onnx_importer.utils.decorators import refactoring_required
 from ngraph_onnx.onnx_importer.utils.matmul import reshape_for_matmul
-from ngraph_onnx.onnx_importer.utils.mapping import onnx_tensor_type_to_numpy_type
+from ngraph_onnx.onnx_importer.utils.types import onnx_tensor_type_to_numpy_type
 from ngraph_onnx.onnx_importer.utils.misc import split_pads_into_pairs
 from ngraph_onnx.onnx_importer.utils.pool import make_pooling_op, make_global_pooling_op
 from ngraph_onnx.onnx_importer.utils.reduction import make_reduction_op, get_reduction_axes
